Move modified CommonMark renderers to own namespace
