+++ app/Libraries/Markdown/CommonMark/Renderers/TableRenderer.php
@@ -3,7 +3,7 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace App\Libraries\Markdown;
+namespace App\Libraries\Markdown\CommonMark\Renderers;
 
 use InvalidArgumentException;
 use League\CommonMark\Block\Element\AbstractBlock;
@@ -12,9 +12,9 @@ use League\CommonMark\ElementRendererInterface;
 use League\CommonMark\Extension\Table\Table;
 use League\CommonMark\HtmlElement;
 
-class OsuTableRenderer implements BlockRendererInterface
+class TableRenderer implements BlockRendererInterface
 {
-    public function render(AbstractBlock $block, ElementRendererInterface $htmlRenderer, $inTightList = false)
+    public function render(AbstractBlock $block, ElementRendererInterface $htmlRenderer, bool $inTightList = false)
     {
         if (!$block instanceof Table) {
             throw new InvalidArgumentException('Incompatible block type: '.get_class($block));
